Add some coloring and temporary vimrc
This adds an initial bit of coloration to the makefile output

It also adds the creation and temporary linkage to a vimrc file
that only includes the required config for installing the Vundle
plugins. This allows us to install all of the Vundle plugins
without any complaints because we're using a vimrc that, for example
tries to use a currently non-existant colorscheme.

Once all of the plugins are installed, we link to the real vimrc.