Fix colors, vundle install, make sections in makefile
This fixes a few issues

Firstly it fixes some color issues. We're now using red for note text
because it stands out more than blue. We also add colors to all of
the text.

This also fixes the Vundle install so that it doesn't display anything
to the user.

Finally, we split the makefile into multiple sections based on what
were doing eg. installing vundle, building ycm etc.